requirements.txt
numpy>=1.21
scipy>=1.7
matplotlib>=3.4
Pillow>=8.0
scikit-image>=0.19
jupyter
pytest
